+++ src/DEFAULT_FORMAT_VALUE_TYPES.mjs
@@ -6,6 +6,7 @@ import { FORMAT_VALUE_TYPE_COLOR, FORMAT_VALUE_TYPE_DATE, FORMAT_VALUE_TYPE_DATE
 /** @typedef {import("./Color/FluxFormatColorValueElement.mjs").FluxFormatColorValueElement} FluxFormatColorValueElement */
 /** @typedef {import("./Url/FluxFormatUrlValueElement.mjs").FluxFormatUrlValueElement} FluxFormatUrlValueElement */
 /** @typedef {import("./formatValue.mjs").formatValue} formatValue */
+/** @typedef {import("./StyleSheetManager/StyleSheetManager.mjs").StyleSheetManager} StyleSheetManager */
 /** @typedef {import("./Time/TimeValue.mjs").TimeValue} TimeValue */
 /** @typedef {import("./Url/UrlValue.mjs").UrlValue} UrlValue */
 
@@ -103,15 +104,17 @@ export async function valueToTimestamp(value = null, string_value_is_local_time_
 
 /**
  * @param {string | null} color
+ * @param {StyleSheetManager | null} style_sheet_manager
  * @returns {Promise<FluxFormatColorValueElement | string | null>}
  */
-export async function formatColorValue(color = null) {
+export async function formatColorValue(color = null, style_sheet_manager = null) {
     if ((color ?? "") === "") {
         return color;
     }
 
     return (await import("./Color/FluxFormatColorValueElement.mjs")).FluxFormatColorValueElement.new(
-        color
+        color,
+        style_sheet_manager
     );
 }
 
@@ -202,14 +205,16 @@ export async function formatTimeValue(time = null) {
 
 /**
  * @param {UrlValue} url
+ * @param {StyleSheetManager | null} style_sheet_manager
  * @returns {Promise<FluxFormatUrlValueElement | string | null>}
  */
-export async function formatUrlValue(url = null) {
+export async function formatUrlValue(url = null, style_sheet_manager = null) {
     if (url === null || typeof url !== "object") {
         return formatUrlValue(
             {
                 url
-            }
+            },
+            style_sheet_manager
         );
     }
 
@@ -220,20 +225,23 @@ export async function formatUrlValue(url = null) {
     return (await import("./Url/FluxFormatUrlValueElement.mjs")).FluxFormatUrlValueElement.new(
         url.url,
         url.label ?? null,
-        url.title ?? null
+        url.title ?? null,
+        style_sheet_manager
     );
 }
 
 /**
  * @param {EmailValue} email
+ * @param {StyleSheetManager | null} style_sheet_manager
  * @returns {Promise<FluxFormatUrlValueElement | string | null>}
  */
-export async function formatEmailValue(email = null) {
+export async function formatEmailValue(email = null, style_sheet_manager = null) {
     if (email === null || typeof email !== "object") {
         return formatEmailValue(
             {
                 email
-            }
+            },
+            style_sheet_manager
         );
     }
 
@@ -246,7 +254,8 @@ export async function formatEmailValue(email = null) {
             url: `mailto:${email.email}`,
             label: (email.label ?? "") !== "" ? email.label : email.email,
             title: email.title ?? null
-        }
+        },
+        style_sheet_manager
     );
 }