+++ src/formatValue.mjs
@@ -1,3 +1,5 @@
+/** @typedef {import("./StyleSheetManager/StyleSheetManager.mjs").StyleSheetManager} StyleSheetManager */
+
 /**
- * @typedef {(value?: *) => Promise<*>} formatValue
+ * @typedef {(value?: *, style_sheet_manager?: StyleSheetManager | null) => Promise<*>} formatValue
  */